+++ CHANGELOG.rst
@@ -7,6 +7,16 @@ Changelog for the **SODAR Core** Django app package. Loosely follows the
 Note that the issue IDs here refer to ones in the private CUBI GitLab.
 
 
+Unreleased
+==========
+
+Removed
+-------
+
+- **Projectroles**
+    - Deprecated ``SODARAPIObjectInProjectPermissions`` base class (#527)
+
+
 v0.8.0 (2020-04-08)
 ===================